# public/openapi.yaml
openapi: 3.1.0
info:
  title: Starfagrunnur API
  description: |
    Open Icelandic occupational knowledge graph. Combines ÍSTARF21 (the Icelandic
    ISCO-08 adaptation) categorisation with ESCO occupations and skills.

    **Data model** (see the graph for the full picture):
    ```
    (OccupationGroup) -[:CHILD_OF]-> (OccupationGroup)     # 4-level ÍSTARF21 hierarchy
    (EscoOccupation) -[:BELONGS_TO_GROUP]-> (OccupationGroup)
    (EscoOccupation) -[:REQUIRES_SKILL {relationType}]-> (Skill)
    (AliasTerm) -[:ALIAS_OF]-> (EscoOccupation)
    ```

    **ÍSTARF21 levels**:
    - Level 1 – bálkur (major group, 1-digit) — 9 categories
    - Level 2 – deild (sub-major, 2-digit) — 39 categories
    - Level 3 – klasi (minor, 3-digit) — 121 categories
    - Level 4 – starfaflokkur (unit, 4-digit) — 409 categories; this is where ESCO roles attach and tasks live

    **Skill classification (ESCO semantics)**:
    - `relationType: essential` — typical for the occupation across the European labour market
    - `relationType: optional` — may be required depending on specialisation/context
    - `skillType: skill/competence` — doing skills
    - `skillType: knowledge` — knowing things

    All labels are in Icelandic (is-IS).
  version: "1.0.0"
  contact:
    name: Peritus
    url: https://starfagrunnur.is
servers:
  - url: /
    description: Current host (the origin serving this documentation)
  - url: https://starfagrunnur.is
    description: Production
...
    description: API alias
...
paths:
...
  /api/v1/occupations/{code}/skills:
    get:
      tags: [Occupations]
      summary: Skills via this group
      description: |
        Skills required by ESCO occupations under this group. Joined via
        (Group)<-[:BELONGS_TO_GROUP]-(Esco)-[:REQUIRES_SKILL]->(Skill).

        By default returns the union across all ESCO roles in the group. Pass
        `occupation=<label>` to filter to one specific ESCO role.

        **Answers questions like:**
        - "What skills does a nurse category (2221) need?"
        - "Show me only *essential* skills for this category." → `?relation_type=essential`
        - "Only list knowledge-type skills, not competencies." → `?skill_type=knowledge`
        - "Show skills *just* for 'hjúkrunarfræðingur', not the whole group." → `?occupation=hjúkrunarfræðingur`
        - For single-ESCO-role queries, `/esco-occupations/{uuid}/skills` is often cleaner (UUID-keyed, no encoding of Icelandic labels).
      parameters:
        - $ref: "#/components/parameters/Code"
        - in: query
          name: relation_type
          schema: { type: string, enum: [essential, optional] }
        - in: query
          name: skill_type
          schema: { type: string, enum: ["skill/competence", "knowledge"] }
        - in: query
          name: occupation
          schema: { type: string }
          description: Filter to skills required by this specific ESCO occupation label (e.g. "hjúkrunarfræðingur").
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Paginated skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillsResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/v1/esco-occupations/{uuid}/skills:
    get:
      tags: [ESCO]
      summary: Skills for this ESCO occupation
      description: |
        Paginated list of skills required by one specific ESCO occupation.

        **Answers questions like:**
        - "What skills does a *hjúkrunarfræðingur* actually need?"
        - "Show me only the essential skills for this role." → `?relation_type=essential`
        - "I want all ~90 skills for this role, paginated." → iterate with `page`
        - Precision-oriented alternative to `/occupations/{code}/skills`, which unions across the whole group.
      parameters:
        - $ref: "#/components/parameters/EscoUuid"
        - in: query
          name: relation_type
          schema: { type: string, enum: [essential, optional] }
        - in: query
          name: skill_type
          schema: { type: string, enum: ["skill/competence", "knowledge"] }
        - in: query
          name: page
          schema: { type: integer, default: 1, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        "200":
          description: Paginated skills
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SkillsResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/skills:
    get:
      tags: [Skills]
      summary: List / search skills
      description: |
        Enumerate ESCO skills. Without `q`, returns skills sorted alphabetically
        (paginated). With `q`, returns ranked full-text matches over Icelandic labels.

        **Answers questions like:**
        - "Is *Python* in the skills database?" → `?q=Python`
        - "List only *knowledge* skills, not competencies." → `?skill_type=knowledge`
        - "Enumerate every skill (paginated)." → no params + paginate
      parameters:
        - in: query
          name: q
          schema: { type: string, minLength: 2 }
          description: Full-text query (fuzzy + wildcard).
        - in: query
          name: skill_type
          schema: { type: string, enum: ["skill/competence", "knowledge"] }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 200 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  limit: { type: integer }
                  offset: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/SkillSummary" }
  /api/v1/skills/{uuid}:
    get:
      tags: [Skills]
      summary: Get a skill
      description: |
        Skill detail: label, description, type, and counts of ESCO occupations
        that require it.

        **Answers questions like:**
        - "What is this skill and what does it actually mean?"
        - "How many occupations require *Python* as essential vs optional?"
        - "Is this a *competence* (doing) or *knowledge* (knowing) skill?"
      parameters:
        - $ref: "#/components/parameters/SkillUuid"
      responses:
        "200":
          description: Skill detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  conceptUri: { type: string, format: uri }
                  uuid: { type: string }
                  preferredLabel_is: { type: string }
                  skillType:
                    type: string
                    enum: ["skill/competence", "knowledge"]
                  description_is: { type: [string, "null"] }
                  occupations:
                    type: object
                    properties:
                      essential_count: { type: integer }
                      optional_count: { type: integer }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    Code:
      name: code
      in: path
      required: true
      schema: { type: string, pattern: "^\\d{1,4}$" }
      description: ÍSTARF21 code (1–4 digits). Length determines level.
      examples:
        balkur:   { value: "2",    summary: "Level 1 — Sérfræðistörf" }
        deild:    { value: "22",   summary: "Level 2 — Sérfræðistörf í heilbrigðisgreinum" }
        klasi:    { value: "222",  summary: "Level 3" }
        unit:     { value: "2221", summary: "Level 4 — Sérfræðistörf við hjúkrun" }
    EscoUuid:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: |
        UUID part of the ESCO conceptUri. The full URI has the form
        `http://data.europa.eu/esco/occupation/<uuid>`; pass just the UUID.
      example: "8d3e8aaa-791b-4c75-a465-f3f827028f50"
    SkillUuid:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
      description: |
        UUID part of the ESCO skill conceptUri. The full URI has the form
        `http://data.europa.eu/esco/skill/<uuid>`; pass just the UUID.
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
  schemas:
...
    SkillSummary:
      type: object
      required: [uuid, conceptUri, preferredLabel_is, skillType]
      properties:
        uuid:              { type: string }
        conceptUri:        { type: string, format: uri }
        preferredLabel_is: { type: string }
        skillType:
          type: string
          enum: ["skill/competence", "knowledge"]
        description_is:    { type: [string, "null"] }
    SkillItem:
      type: object
      required: [preferredLabel_is, skillType, relationType]
      properties:
        preferredLabel_is: { type: string }
        skillType:
          type: string
          enum: ["skill/competence", "knowledge"]
        description_is: { type: [string, "null"] }
        relationType:
          type: string
          enum: [essential, optional]
        occupation:
          type: string
          description: Present only on the group-level /skills endpoint — names the ESCO role requiring the skill.
    SkillsResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SkillItem" }
        total:       { type: integer }
        page:        { type: integer }
        page_size:   { type: integer }
        total_pages: { type: integer }
...